requirements.txt
# Python dependencies
flask>=2.3.0
flask-cors>=4.0.0
flask-sqlalchemy>=3.0.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
flask-socketio>=5.3.6
python-socketio[client]>=5.11.0
eventlet>=0.35.0

# Process inspection
psutil>=5.9.0

# Security dependencies
bcrypt>=4.0.0
flask-limiter>=3.0.0
werkzeug>=3.0.0
PyJWT>=2.8.0

# (Removed AI provider packages to slim environment; restore if reinstating AI assistant)
